:recycle: refactor: organiza a lógica de negócio do prismaQuizRepository em um novo arquivo
